extract review_notification
